[BC]Refactor #66:  Refactor tests for Escrow contract deployment and deposit functionality with ERC20Permit
